Make Scv Storage
scv 저장공간과 함수를 분해하여 각각 저장클래스와 그 클래스를 다루는 함수로 분해 중
