+++ WorkSpace/ASB/ASB.vcxproj.filters
@@ -8,10 +8,14 @@
     <ClCompile Include="game.cpp" />
     <ClCompile Include="LiveObject\Enemy.cpp" />
     <ClCompile Include="LiveObject\Player.cpp" />
-    <ClCompile Include="DataStorage\CsvDataStorage.cpp" />
-    <ClCompile Include="DataStorage\BaseData.cpp" />
-    <ClCompile Include="DataStorage\DataClass\AllNodePattenClass.cpp" />
-    <ClCompile Include="DataStorage\DataClass\EnemyAtkPattenData.cpp" />
+    <ClCompile Include="CsvData\CsvDataManager.cpp" />
+    <ClCompile Include="CsvData\BaseData.cpp" />
+    <ClCompile Include="CsvData\DataClass\AllNodePattenClass.cpp" />
+    <ClCompile Include="CsvData\DataClass\EnemyAtkPattenData.cpp" />
+    <ClCompile Include="CsvData\DataStorage\CsvDataStorage.cpp" />
+    <ClCompile Include="CsvData\DataClass\EnemyData.cpp" />
+    <ClCompile Include="CsvData\DataClass\PlayerAtkPetternData.cpp" />
+    <ClCompile Include="CsvData\DataClass\PlayerData.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GameLogicTestScene\GameLogicTestScene.h" />
@@ -19,9 +23,13 @@
     <ClInclude Include="LogicTestApplication\LogicTestApplication.h" />
     <ClInclude Include="LiveObject\Enemy.h" />
     <ClInclude Include="LiveObject\Player.h" />
-    <ClInclude Include="DataStorage\CsvDataStorage.h" />
-    <ClInclude Include="DataStorage\BaseData.h" />
-    <ClInclude Include="DataStorage\DataClass\AllNodePattenClass.h" />
-    <ClInclude Include="DataStorage\DataClass\EnemyAtkPattenData.h" />
+    <ClInclude Include="CsvData\CsvDataManager.h" />
+    <ClInclude Include="CsvData\BaseData.h" />
+    <ClInclude Include="CsvData\DataClass\AllNodePattenClass.h" />
+    <ClInclude Include="CsvData\DataClass\EnemyAtkPattenData.h" />
+    <ClInclude Include="CsvData\DataStorage\CsvDataStorage.h" />
+    <ClInclude Include="CsvData\DataClass\EnemyData.h" />
+    <ClInclude Include="CsvData\DataClass\PlayerAtkPetternData.h" />
+    <ClInclude Include="CsvData\DataClass\PlayerData.h" />
   </ItemGroup>
 </Project>
