+++ WorkSpace/ASB/CsvData/DataClass/AllNodePattenClass.cpp
