+++ WorkSpace/ASB/CsvData/DataClass/EnemyData.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "../BaseData.h"
+
+
+class EnemyData : public BaseData
+{
+};
+
