+++ WorkSpace/ASB/CsvData/DataClass/PlayerAtkPetternData.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "../BaseData.h"
+
+class PlayerAtkPetternData : public BaseData
+{
+};
+
