+++ WorkSpace/ASB/CsvData/DataClass/PlayerData.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "../BaseData.h"
+
+class PlayerData : public BaseData
+{
+
+};
+
